feat: 로그인 한 회원 정보 UI에 표시 #24
